# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  models.CreateTaskInput:
    properties:
      completed:
        type: boolean
      description:
        type: string
      title:
        type: string
    type: object
  models.Task:
    properties:
      completed:
        type: boolean
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.UpdateTaskInput:
    properties:
      completed:
        type: boolean
      description:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
  description: REST API для управления задачами
  title: Tasks API
  version: "1.0"
paths:
  /tasks:
    get:
      description: Возвращает массив всех задач из базы данных
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Task'
            type: array
        "500":
          description: Unable to fetch tasks
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список всех задач
      tags:
      - tasks
  /tasks/{id}:
    delete:
      description: Удаляет задачу по её ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid task ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить задачу
      tags:
      - tasks
    get:
      description: Возвращает одну задачу по её идентификатору
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Invalid task ID
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить задачу по ID
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Обновляет поля задачи по её ID
      parameters:
      - description: ID задачи
        in: path
        name: id
        required: true
        type: integer
      - description: Поля для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTaskInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Task not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить задачу
      tags:
      - tasks
  /tasks/create:
    post:
      consumes:
      - application/json
      description: Создаёт задачу на основе переданного JSON
      parameters:
      - description: Данные для создания задачи
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateTaskInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Task'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать новую задачу
      tags:
      - tasks
swagger: "2.0"